+++ branches/newui/AnnotationDisplay.cpp
@@ -81,7 +81,7 @@ void AnnotationDisplay::displayTrack(const vector<track_entry>& track)
 	for(int row = 0; row < display_size / width; row++)
 	{
 		//check to see if any of the old tracks are done
-		for(int k = 0; k < activeEntries.size(); ++k)
+        for(int k = 0; k < (int)activeEntries.size(); ++k)
 		{
 			if( !activeEntries[k].isBlank() )
 			{
@@ -95,9 +95,9 @@ void AnnotationDisplay::displayTrack(const vector<track_entry>& track)
 			}
 		}
 		//check to match start for a new track
-		while(next_spot < track.size() && track[next_spot].stop < pix_start )//assumes tracks are in order
+        while(next_spot < (int)track.size() && track[next_spot].stop < pix_start )//assumes tracks are in order
 			next_spot++;
-		while(next_spot < track.size() 
+        while(next_spot < (int)track.size()
 			&& ((track[next_spot].start >= pix_start && track[next_spot].start <= pix_stop)//start in range
 			    || (track[next_spot].stop >= pix_start && track[next_spot].stop <= pix_stop)//end in range
 				|| (track[next_spot].start < pix_start && track[next_spot].stop > pix_stop)) )//in the middle
@@ -108,7 +108,7 @@ void AnnotationDisplay::displayTrack(const vector<track_entry>& track)
 		//display each track
 		glPushMatrix();
 			glScaled(1,-1,1);
-		for(int x = 0; x < activeEntries.size(); ++x)
+        for(int x = 0; x < (int)activeEntries.size(); ++x)
 		{
 			color c;
 			if(activeEntries[x].isBlank())
@@ -131,7 +131,7 @@ void AnnotationDisplay::displayTrack(const vector<track_entry>& track)
 
 void AnnotationDisplay::stackEntry(vector<track_entry>& activeEntries, track_entry item)
 {
-		for(int k = 0; k < activeEntries.size(); ++k)
+        for(int k = 0; k < (int)activeEntries.size(); ++k)
 		{
 			if( activeEntries[k].isBlank() )
 			{
@@ -152,7 +152,7 @@ string AnnotationDisplay::mouseClick(point2D pt)
 		int stop = start + ui->widthDial->value();
 		if(!gtfTrack.empty())
 		{
-			for(int i = 0; i < gtfTrack.size(); ++i)
+            for(int i = 0; i < (int)gtfTrack.size(); ++i)
 			{
 				if(((gtfTrack[i].start >= start && gtfTrack[i].start <= stop)//start in range
 				    || (gtfTrack[i].stop >= start && gtfTrack[i].stop <= stop)//end in range
